feat: 리뷰 이미지 모달 구현
Resolves: #320
