--- frontend/src/components/ImageModal/ImageModal.styled.ts
@@ -1,33 +1,0 @@
-import styled from '@emotion/styled';
-
-const Container = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  left: 0;
-  overflow: hidden;
-  z-index: 1;
-`;
-
-const Backdrop = styled.div`
-  width: 100%;
-  height: 100vh;
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 9999;
-  background-color: rgba(0, 0, 0, 0.5);
-`;
-
-const Img = styled.img`
-  min-width: 600px;
-  max-width: 900px;
-  height: auto;
-  z-index: 10000;
-`;
-
-export { Container, Img, Backdrop };
